--- packages/server/src/main/java/software/uncharted/terarium/hmiserver/models/dataservice/GroupPermissions.java
@@ -1,11 +1,0 @@
-package software.uncharted.terarium.hmiserver.models.dataservice;
-
-import software.uncharted.terarium.hmiserver.models.dataservice.permission.PermissionGroup;
-import software.uncharted.terarium.hmiserver.models.dataservice.permission.PermissionUser;
-
-import java.util.List;
-
-public class GroupPermissions {
-	private List<PermissionGroup> permissionGroups;
-	private List<PermissionUser> permissionUsers;
-}
