+++ packages/server/src/main/java/software/uncharted/terarium/hmiserver/models/permissions/PermissionRelationships.java
@@ -1,21 +1,25 @@
-package software.uncharted.terarium.hmiserver.models.dataservice.permission;
+package software.uncharted.terarium.hmiserver.models.permissions;
 
+import software.uncharted.terarium.hmiserver.annotations.TSModel;
 import software.uncharted.terarium.hmiserver.utils.rebac.Schema;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@TSModel
 public class PermissionRelationships {
 	private List<PermissionGroup> permissionGroups = new ArrayList<>();
 	private List<PermissionUser> permissionUsers = new ArrayList<>();
 	private List<PermissionProject> permissionProjects = new ArrayList<>();
 
-	public void addUser(String id, Schema.Relationship relationship) {
-		permissionUsers.add(new PermissionUser(id, relationship.toString()));
+	public void addUser(PermissionUser permissionUser, Schema.Relationship relationship) {
+		permissionUser.setRelationship(relationship.toString());
+		permissionUsers.add(permissionUser);
 	}
 
-	public void addGroup(String id, Schema.Relationship relationship) {
-		permissionGroups.add(new PermissionGroup(id, relationship.toString()));
+	public void addGroup(PermissionGroup permissionGroup, Schema.Relationship relationship) {
+		permissionGroup.setRelationship(relationship.toString());
+		permissionGroups.add(permissionGroup);
 	}
 
 	public void addProject(String id, Schema.Relationship relationship) {
